use default deserializer config

+++ api/bindings_compiler.h
@@ -140,11 +140,11 @@ typedef struct {
   /**
    * Compiler version. set 0 to unset and to use default
    */
-  uint8_t compiler_version;
+  uint32_t compiler_version;
   /**
    * language version. set 0 to unset and to use default
    */
-  uint8_t language_version;
+  uint32_t language_version;
 } CompilerBuildConfig;
 
 typedef struct {
